genai-project/backend/requirements.txt
Flask
pymongo
python-dotenv
google-generativeai
